refactor: 수익 계산 책임 Outcome으로 위임
