+++ src/main/java/domain/participant/BettingMoney.java
@@ -1,6 +1,5 @@
 package domain.participant;
 
-import domain.game.Outcome;
 import java.util.Objects;
 
 public class BettingMoney {
@@ -15,10 +14,6 @@ public class BettingMoney {
         if (amount <= 0) {
             throw new IllegalArgumentException("베팅 금액이 0이하 일 수 없습니다.");
         }
-    }
-
-    public int calculateProfit(Outcome outcome) {
-        return (int) (amount * outcome.getProfitRate());
     }
 
     public int getAmount() {
